Initial stage of import/export. No support yet for JSON/CSV yet. Also, Import does not support files

+++ resources/views/records/fieldInputs/gallery.blade.php
@@ -9,14 +9,18 @@ if(file_exists($dir)) {
             unlink($dir.'/'.$file->getFilename());
         }
     }
-    foreach (new \DirectoryIterator($dir.'/thumbnail') as $file) {
-        if ($file->isFile()) {
-            unlink($dir.'/thumbnail/'.$file->getFilename());
+    if(file_exists($dir.'/thmubnail')) {
+        foreach (new \DirectoryIterator($dir.'/thumbnail') as $file) {
+            if ($file->isFile()) {
+                unlink($dir.'/thumbnail/'.$file->getFilename());
+            }
         }
     }
-    foreach (new \DirectoryIterator($dir.'/medium') as $file) {
-        if ($file->isFile()) {
-            unlink($dir.'/medium/'.$file->getFilename());
+    if(file_exists($dir.'/medium')) {
+        foreach (new \DirectoryIterator($dir.'/medium') as $file) {
+            if ($file->isFile()) {
+                unlink($dir.'/medium/'.$file->getFilename());
+            }
         }
     }
 }
